Split ConfigurationManager responsibilities, quality of life refactor
