+++ src/main/groovy/com/blackducksoftware/integration/hub/artifactory/RestTemplateContainer.java
@@ -50,8 +50,6 @@ import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriTemplateHandler;
 
-import com.blackducksoftware.integration.exception.EncryptionException;
-
 @Component
 class RestTemplateContainer extends RestTemplate {
     @Autowired
@@ -60,9 +58,8 @@ class RestTemplateContainer extends RestTemplate {
     private RestTemplate restTemplate;
 
     @PostConstruct
-    public void init() throws EncryptionException {
+    public void init() {
         restTemplate = new RestTemplate();
-
         final String artifactoryUsername = configurationProperties.getArtifactoryUsername();
         final String artifactoryApiKey = configurationProperties.getArtifactoryApiKey();
         if (StringUtils.isNotBlank(artifactoryUsername) && StringUtils.isNotBlank(artifactoryApiKey)) {
@@ -142,105 +139,102 @@ class RestTemplateContainer extends RestTemplate {
     }
 
     @Override
-    public <T> T getForObject(final String url, final Class<T> responseType, final Object... uriVariables) throws RestClientException {
+    public <T> T getForObject(final String url, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.getForObject(url, responseType, uriVariables);
     }
 
     @Override
-    public <T> T getForObject(final String url, final Class<T> responseType, final Map<String, ?> uriVariables) throws RestClientException {
+    public <T> T getForObject(final String url, final Class<T> responseType, final Map<String, ?> uriVariables) {
         return restTemplate.getForObject(url, responseType, uriVariables);
     }
 
     @Override
-    public <T> T getForObject(final URI url, final Class<T> responseType) throws RestClientException {
+    public <T> T getForObject(final URI url, final Class<T> responseType) {
         return restTemplate.getForObject(url, responseType);
     }
 
     @Override
-    public <T> ResponseEntity<T> getForEntity(final String url, final Class<T> responseType, final Object... uriVariables) throws RestClientException {
+    public <T> ResponseEntity<T> getForEntity(final String url, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.getForEntity(url, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> getForEntity(final String url, final Class<T> responseType, final Map<String, ?> uriVariables) throws RestClientException {
+    public <T> ResponseEntity<T> getForEntity(final String url, final Class<T> responseType, final Map<String, ?> uriVariables) {
         return restTemplate.getForEntity(url, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> getForEntity(final URI url, final Class<T> responseType) throws RestClientException {
+    public <T> ResponseEntity<T> getForEntity(final URI url, final Class<T> responseType) {
         return restTemplate.getForEntity(url, responseType);
     }
 
     @Override
-    public HttpHeaders headForHeaders(final String url, final Object... uriVariables) throws RestClientException {
+    public HttpHeaders headForHeaders(final String url, final Object... uriVariables) {
         return restTemplate.headForHeaders(url, uriVariables);
     }
 
     @Override
-    public HttpHeaders headForHeaders(final String url, final Map<String, ?> uriVariables) throws RestClientException {
+    public HttpHeaders headForHeaders(final String url, final Map<String, ?> uriVariables) {
         return restTemplate.headForHeaders(url, uriVariables);
     }
 
     @Override
-    public HttpHeaders headForHeaders(final URI url) throws RestClientException {
+    public HttpHeaders headForHeaders(final URI url) {
         return restTemplate.headForHeaders(url);
     }
 
     @Override
-    public URI postForLocation(final String url, final Object request, final Object... uriVariables) throws RestClientException {
+    public URI postForLocation(final String url, final Object request, final Object... uriVariables) {
         return restTemplate.postForLocation(url, request, uriVariables);
     }
 
     @Override
-    public URI postForLocation(final String url, final Object request, final Map<String, ?> uriVariables) throws RestClientException {
+    public URI postForLocation(final String url, final Object request, final Map<String, ?> uriVariables) {
         return restTemplate.postForLocation(url, request, uriVariables);
     }
 
     @Override
-    public URI postForLocation(final URI url, final Object request) throws RestClientException {
+    public URI postForLocation(final URI url, final Object request) {
         return restTemplate.postForLocation(url, request);
     }
 
     @Override
-    public <T> T postForObject(final String url, final Object request, final Class<T> responseType, final Object... uriVariables) throws RestClientException {
+    public <T> T postForObject(final String url, final Object request, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.postForObject(url, request, responseType, uriVariables);
     }
 
     @Override
-    public <T> T postForObject(final String url, final Object request, final Class<T> responseType, final Map<String, ?> uriVariables)
-            throws RestClientException {
+    public <T> T postForObject(final String url, final Object request, final Class<T> responseType, final Map<String, ?> uriVariables) {
         return restTemplate.postForObject(url, request, responseType, uriVariables);
     }
 
     @Override
-    public <T> T postForObject(final URI url, final Object request, final Class<T> responseType) throws RestClientException {
+    public <T> T postForObject(final URI url, final Object request, final Class<T> responseType) {
         return restTemplate.postForObject(url, request, responseType);
     }
 
     @Override
-    public <T> ResponseEntity<T> postForEntity(final String url, final Object request, final Class<T> responseType, final Object... uriVariables)
-            throws RestClientException {
+    public <T> ResponseEntity<T> postForEntity(final String url, final Object request, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.postForEntity(url, request, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> postForEntity(final String url, final Object request, final Class<T> responseType, final Map<String, ?> uriVariables)
-            throws RestClientException {
+    public <T> ResponseEntity<T> postForEntity(final String url, final Object request, final Class<T> responseType, final Map<String, ?> uriVariables) {
         return restTemplate.postForEntity(url, request, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> postForEntity(final URI url, final Object request, final Class<T> responseType) throws RestClientException {
+    public <T> ResponseEntity<T> postForEntity(final URI url, final Object request, final Class<T> responseType) {
         return restTemplate.postForEntity(url, request, responseType);
     }
 
     @Override
-    public void put(final String url, final Object request, final Object... uriVariables) throws RestClientException {
+    public void put(final String url, final Object request, final Object... uriVariables) {
         restTemplate.put(url, request, uriVariables);
     }
 
     @Override
-    public void put(final String url, final Object request, final Map<String, ?> uriVariables) throws RestClientException {
+    public void put(final String url, final Object request, final Map<String, ?> uriVariables) {
         restTemplate.put(url, request, uriVariables);
     }
 
@@ -250,7 +244,7 @@ class RestTemplateContainer extends RestTemplate {
     }
 
     @Override
-    public <T> T patchForObject(final String url, final Object request, final Class<T> responseType, final Object... uriVariables) throws RestClientException {
+    public <T> T patchForObject(final String url, final Object request, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.patchForObject(url, request, responseType, uriVariables);
     }
 
@@ -261,108 +255,94 @@ class RestTemplateContainer extends RestTemplate {
     }
 
     @Override
-    public <T> T patchForObject(final URI url, final Object request, final Class<T> responseType) throws RestClientException {
+    public <T> T patchForObject(final URI url, final Object request, final Class<T> responseType) {
         return restTemplate.patchForObject(url, request, responseType);
     }
 
     @Override
-    public void delete(final String url, final Object... uriVariables) throws RestClientException {
+    public void delete(final String url, final Object... uriVariables) {
         restTemplate.delete(url, uriVariables);
     }
 
     @Override
-    public void delete(final String url, final Map<String, ?> uriVariables) throws RestClientException {
+    public void delete(final String url, final Map<String, ?> uriVariables) {
         restTemplate.delete(url, uriVariables);
     }
 
     @Override
-    public void delete(final URI url) throws RestClientException {
+    public void delete(final URI url) {
         restTemplate.delete(url);
     }
 
     @Override
-    public Set<HttpMethod> optionsForAllow(final String url, final Object... uriVariables) throws RestClientException {
+    public Set<HttpMethod> optionsForAllow(final String url, final Object... uriVariables) {
         return restTemplate.optionsForAllow(url, uriVariables);
     }
 
     @Override
-    public Set<HttpMethod> optionsForAllow(final String url, final Map<String, ?> uriVariables) throws RestClientException {
+    public Set<HttpMethod> optionsForAllow(final String url, final Map<String, ?> uriVariables) {
         return restTemplate.optionsForAllow(url, uriVariables);
     }
 
     @Override
-    public Set<HttpMethod> optionsForAllow(final URI url) throws RestClientException {
+    public Set<HttpMethod> optionsForAllow(final URI url) {
         return restTemplate.optionsForAllow(url);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity, final Class<T> responseType,
-            final Object... uriVariables)
-                    throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity, final Class<T> responseType, final Object... uriVariables) {
         return restTemplate.exchange(url, method, requestEntity, responseType, uriVariables);
     }
 
     @Override
     public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity, final Class<T> responseType,
             final Map<String, ?> uriVariables)
-                    throws RestClientException {
+    {
         return restTemplate.exchange(url, method, requestEntity, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final URI url, final HttpMethod method, final HttpEntity<?> requestEntity, final Class<T> responseType)
-            throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final URI url, final HttpMethod method, final HttpEntity<?> requestEntity, final Class<T> responseType) {
         return restTemplate.exchange(url, method, requestEntity, responseType);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity,
-            final ParameterizedTypeReference<T> responseType,
-            final Object... uriVariables) throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity, final ParameterizedTypeReference<T> responseType, final Object... uriVariables) {
         return restTemplate.exchange(url, method, requestEntity, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity,
-            final ParameterizedTypeReference<T> responseType,
-            final Map<String, ?> uriVariables) throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final String url, final HttpMethod method, final HttpEntity<?> requestEntity, final ParameterizedTypeReference<T> responseType, final Map<String, ?> uriVariables) {
         return restTemplate.exchange(url, method, requestEntity, responseType, uriVariables);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final URI url, final HttpMethod method, final HttpEntity<?> requestEntity,
-            final ParameterizedTypeReference<T> responseType)
-                    throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final URI url, final HttpMethod method, final HttpEntity<?> requestEntity, final ParameterizedTypeReference<T> responseType) {
         return restTemplate.exchange(url, method, requestEntity, responseType);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final RequestEntity<?> requestEntity, final Class<T> responseType) throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final RequestEntity<?> requestEntity, final Class<T> responseType) {
         return restTemplate.exchange(requestEntity, responseType);
     }
 
     @Override
-    public <T> ResponseEntity<T> exchange(final RequestEntity<?> requestEntity, final ParameterizedTypeReference<T> responseType) throws RestClientException {
+    public <T> ResponseEntity<T> exchange(final RequestEntity<?> requestEntity, final ParameterizedTypeReference<T> responseType) {
         return restTemplate.exchange(requestEntity, responseType);
     }
 
     @Override
-    public <T> T execute(final String url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor,
-            final Object... uriVariables)
-                    throws RestClientException {
+    public <T> T execute(final String url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor, final Object... uriVariables) {
         return restTemplate.execute(url, method, requestCallback, responseExtractor, uriVariables);
     }
 
     @Override
-    public <T> T execute(final String url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor,
-            final Map<String, ?> uriVariables)
-                    throws RestClientException {
+    public <T> T execute(final String url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor, final Map<String, ?> uriVariables) {
         return restTemplate.execute(url, method, requestCallback, responseExtractor, uriVariables);
     }
 
     @Override
-    public <T> T execute(final URI url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor)
-            throws RestClientException {
+    public <T> T execute(final URI url, final HttpMethod method, final RequestCallback requestCallback, final ResponseExtractor<T> responseExtractor) {
         return restTemplate.execute(url, method, requestCallback, responseExtractor);
     }
 }
